[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int strange(int a, int b)
+long long strange(long long a, long long b)
 {
     return (a + b) * (a - b);
 }
@@ -14,7 +14,7 @@ int main() {
     int a, b;
     cin >> a >> b;
 
-    int c = 0;
+    long long c = 0;
     c = strange(a, b);
     
     cout << c << "\n";
